Added code to handle the getApplication method of the Activity class.
LibraryClassPatcher: Added code to create an ApplicationHolder class, and patch the Activity::getApplication method to return the Application it holds.
AndroidEntryPointCreator: Added code to initialize the static field in the ApplicationHolder with the singleton Application instance.
EasyTaintWrapper: changed the 'excludeTaints' option in the hasWrappedMethodsForClass method to false (It doesn't find the flow otherwise)